Change map_bits to return Vec rather than iterator
bit_data::map_bits returns an iterator, but internally, it first constructs a
vector. All consumers immediately collect into a vector. The whole thing becomes
a bit simpler if we don't use the intermediate iterator.

I spent a bit of time trying to make a clear implementation of map_bits that
does not use the intermediate vector, but finally stopped this effort in favor of
the present. If there were other consumers that could use an iterator, it might
be worth taking this up again. But at present, there are none.